refactor(split): Split into multiple classes
